fix: wrap GalleryItemChart component
